update-to-work-version-of-unoconv
While debugging the error we found that the command given my devs does nto work
because devs uses the latest version of unoconv from github.

This allow us to pull that script and set as a specific version of the script
into teh base image to allow tesing. Run manually using this script does work,
the only bit left is run thorugh dotnet which needs to be tested.